% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Error.R
\name{Error}
\alias{Error}
\title{Error}
\usage{
Error(l, prop_case, beta)
}
\arguments{
\item{l}{matrix containing outcome and covariates from test set}

\item{prop_case}{proportion of cases in the training set}

\item{beta}{coefficients of logistic model}
}
\value{

}
\description{
Error
}
