% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Logistic.R
\name{Logistic}
\alias{Logistic}
\title{Logistic}
\usage{
Logistic(l)
}
\arguments{
\item{l}{a matrix containing outcome and markers. Please put outcome in the
first column of the matrix, and put markers in the rest columns}
}
\value{
it returns the MLE of coefficients while fixing the first component
of coefficents to be 1.
}
\description{
This a function to get MLE of a logitic model while fixing the first component
of coefficents to be 1.
}
