% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SAUC.R
\name{SAUC}
\alias{SAUC}
\title{SAUC}
\usage{
SAUC(l)
}
\arguments{
\item{l}{a matrix containing outcome and markers. Please put outcome in the
first column of the matrix, and put markers in the rest columns}
}
\value{
it returns the SAUC estimates of coefficients while fixing the first component
of coefficents to be 1.
}
\description{
This function gets SAUC estimates of a logistic model while fixing the first component
of coefficents to be 1.
}
