% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SimData1_100.R
\docType{data}
\name{SimData1_100}
\alias{SimData1_100}
\title{Things manatees tell you to lift your spirits}
\format{
a list with 2 components
}
\source{
sampled from certain distribution
}
\usage{
SimData1_100
}
\description{
A dataset containing several manatee quotes
}
\keyword{datasets}
