% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SimSAUC.R
\name{SimSAUC}
\alias{SimSAUC}
\title{SimSAUC}
\usage{
SimSAUC(y, x, sim_size, parallel = F, n_core = 1, ratio = 2/3, seed = 2021)
}
\arguments{
\item{y}{binary outcome of a logistic model}

\item{x}{a dataframe containing markers}

\item{sim_size}{simulation size}

\item{parallel}{TRUE if parallel computation is applied}

\item{n_core}{the numbe of cores used in parallel computation}

\item{ratio}{threshold for decsion rule}

\item{seed}{random seed for reproducibility}
}
\value{
'beta_SAUC mean' is the SAUC estimates from simulations.
'beta_Logistic' is the mean MLE from simulations. 'sd_of_beta_SAUC' is the
standard error of SAUC estimates. 'sd_of_beta_Logistic' is the standard error
of SAUC estimates. 'error_SAUC' is mean error of SAUC and 'error_Logistic'
is mean error of MLE. 'sd_of_error_SAUC' and 'sd_of_error_Logistic' are
standard errors of mean errors of SAUC and MLE.
}
\description{
This is a function to automatically run simulations given a simulated data.
This function will split data into training set and testing set for a given times.
It then computed SAUC estimator and MLE in training set. And using the estimated model,
function computed error rate in the testing set.
}
