% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Error.R
\name{Error}
\alias{Error}
\title{Error}
\usage{
Error(l, prop_case, beta)
}
\arguments{
\item{l}{a matrix containing outcome and markers. Please put outcome in the
first column of the matrix, and put markers in the rest columns}

\item{prop_case}{threshold for decsion rule}

\item{beta}{coefficients of logistic model}
}
\value{
it returns the error rate which is the proportion of falsely
classified subjects in the given data.
}
\description{
This function calculate error rate of a logistic model in the given data.
}
